Time: 4 ms (10.11%), Space: 45.1 MB (44.76%) - LeetHub

+++ 1423-maximum-points-you-can-obtain-from-cards/1423-maximum-points-you-can-obtain-from-cards.cpp
@@ -1,23 +1,25 @@
 class Solution {
 public:
     int maxScore(vector<int>& cardPoints, int k) {
-        long long lsum = 0, rsum = 0, mSum = 0;
-
-        // Calculate the sum of the first k cards
+        
+        int lsum = 0, rsum = 0, mSum = 0; // Initialize variables
+        
+        // Calculate the left sum and initial max sum
         for (int i = 0; i < k; i++) {
             lsum += cardPoints[i];
         }
         mSum = lsum;
 
-        // Shift the window to include cards from the right
         int rInd = cardPoints.size() - 1;
+
+        // Adjust left sum and calculate right sum
         for (int i = k - 1; i >= 0; i--) {
             lsum -= cardPoints[i];
             rsum += cardPoints[rInd];
             rInd--;
-            mSum = max(mSum, lsum + rsum);
+            mSum = max(mSum, lsum + rsum); // Update max sum
         }
 
-        return static_cast<int>(mSum);
+        return mSum;
     }
 };
